Split `Pact.Analyze.Types`.
