+++ src/Pact/Analyze/Check.hs
@@ -74,7 +74,6 @@ import qualified Pact.Types.Typecheck      as TC
 import           Pact.Analyze.Analyze      hiding (invariants)
 import           Pact.Analyze.Errors
 import           Pact.Analyze.Parse        (expToCheck, expToInvariant)
-import           Pact.Analyze.Term         (ETerm)
 import           Pact.Analyze.Translate
 import           Pact.Analyze.Types
 import           Pact.Analyze.Util
