+++ tests/AnalyzeSpec.hs
@@ -31,7 +31,6 @@ import           Pact.Types.Runtime           (ModuleData, eeRefStore,
                                                rsModules)
 
 import           Pact.Analyze.Check
-import           Pact.Analyze.Numerical
 import           Pact.Analyze.Parse           (TableEnv, expToProp, inferProp)
 import           Pact.Analyze.PrenexNormalize (prenexConvert)
 import           Pact.Analyze.Types
